Migrate fetch to TS
